Update CheckoutScreen.py
Various updates to the layout of the Checkout screen, such as the retriever essential logo in the top right corner and changes to the display of the items in the cart. Also added was the ability to auto-load the cart with information from a database that holds information about what items are in the cart. In the "For Testing purposes only" portion, this will display 3 items with a random image. These will be removed once functionality to take in data from the database is incorporated.