added sql/crud support to item_class
Added backend SQL support within the Item class. When a new item object is created, it uses the crud.py add item function to create this item in the SQL database. When an item object is deleted using the class's destructor, the object is removed from the SQL database using crud.py remove item function.