Update and rename StudentInventoryView_1.0.py to StudentInventoryView.py
Changed how the student inventory view layout works. Instead of using pages to view all the items, a scroll bar is introduced when the objects exceed the screen limit. Additionally, it uses a standard item from CustomObj named StudentInventoryView that will allow additional objects to be added. The images and labels for the objects will be changed based on the input taken from the database.